mod: put token for ingredient_category endpoint

+++ tests/integration/ingredient_category_test.py
@@ -4,8 +4,10 @@ from fastapi.testclient import TestClient
 url_prefix = '/ingredient_categories'
 ADMIN_ID = "db67b3f4-0e04-47bb-bc46-94826847ee4f"
 
-def test_get_ingredient_category_list(client: TestClient):
-    response = client.get(f"{url_prefix}/")
+def test_get_ingredient_category_list(client: TestClient, token: str):
+    response = client.get(f"{url_prefix}/", 
+        headers={"Authorization": f"Bearer {token}"}
+    )
 
     assert response.status_code == 200
     assert response.json()["detail"] == "Ingredient Category list is retrieved successfully"
@@ -27,8 +29,24 @@ def test_get_ingredient_category_list(client: TestClient):
     assert "updated_date" in ingredient_categories[1]
     assert "created_date" in ingredient_categories[1]
 
-def test_get_ingredient_category_by_id(client: TestClient):
-    response = client.get(f"{url_prefix}/b4b165f6-a4f2-45f6-bda6-0a49092d3f03")
+def test_get_ingredient_category_list_with_invalid_token(client: TestClient):
+    response = client.get(f"{url_prefix}/", 
+        headers={"Authorization": f"Bearer invalid_token"}
+    )
+
+    assert response.status_code == 401
+    assert response.json()["detail"] == "Invalid token"
+
+def test_get_ingredient_category_list_without_token(client: TestClient):
+    response = client.get(f"{url_prefix}/")
+
+    assert response.status_code == 401
+    assert response.json()["detail"] == "Not authenticated"
+
+def test_get_ingredient_category_by_id(client: TestClient, token: str):
+    response = client.get(f"{url_prefix}/b4b165f6-a4f2-45f6-bda6-0a49092d3f03", 
+        headers={"Authorization": f"Bearer {token}"}
+    )
 
     assert response.status_code == 200
     assert response.json()["detail"] == f"Id b4b165f6-a4f2-45f6-bda6-0a49092d3f03 as Ingredient Category is retrieved successfully"
@@ -41,17 +59,36 @@ def test_get_ingredient_category_by_id(client: TestClient):
     assert "updated_date" in ingredient_category
     assert "created_date" in ingredient_category
 
-def test_get_ingredient_category_by_wrong_id(client: TestClient):
-    response = client.get(f"{url_prefix}/b4b165f6-a4f2-45f6-bda6-0a49092d3f01")
+def test_get_ingredient_category_by_id_with_invalid_token(client: TestClient):
+    response = client.get(f"{url_prefix}/b4b165f6-a4f2-45f6-bda6-0a49092d3f03", 
+        headers={"Authorization": f"Bearer invalid_token"}
+    )
+
+    assert response.status_code == 401
+    assert response.json()["detail"] == "Invalid token"
+
+def test_get_ingredient_category_by_id_without_token(client: TestClient):
+    response = client.get(f"{url_prefix}/b4b165f6-a4f2-45f6-bda6-0a49092d3f03")
+
+    assert response.status_code == 401
+    assert response.json()["detail"] == "Not authenticated"
+
+def test_get_ingredient_category_by_wrong_id(client: TestClient, token: str):
+    response = client.get(f"{url_prefix}/b4b165f6-a4f2-45f6-bda6-0a49092d3f01", 
+        headers={"Authorization": f"Bearer {token}"}
+    )
 
     assert response.status_code == 404
     assert response.json()['detail'] == "Id b4b165f6-a4f2-45f6-bda6-0a49092d3f01 as Ingredient Category is not found"
 
-def test_post_ingredient_category(client: TestClient):
-    response = client.post(f"{url_prefix}/", json={
-        "name": "test ingredient category",
-        "created_by": ADMIN_ID
-    })
+def test_post_ingredient_category(client: TestClient, token: str):
+    response = client.post(f"{url_prefix}/", 
+        json={
+            "name": "test ingredient category",
+            "created_by": ADMIN_ID
+        }, 
+        headers={"Authorization": f"Bearer {token}"}
+    )
 
     assert response.status_code == 201
     assert response.json()["detail"] == f"test ingredient category as Ingredient Category is successfully created"
@@ -63,20 +100,68 @@ def test_post_ingredient_category(client: TestClient):
     assert "updated_date" in ingredient_category
     assert "created_date" in ingredient_category
 
-def test_post_ingredient_category_with_duplicate_name(client: TestClient):
-    response = client.post(f"{url_prefix}/", json={
-        "name": "proteins",
-        "created_by": ADMIN_ID
-    })
+def test_post_ingredient_category_with_invalid_token(client: TestClient):
+    response = client.post(f"{url_prefix}/", 
+        json={
+            "name": "test ingredient category",
+            "created_by": ADMIN_ID
+        }, 
+        headers={"Authorization": f"Bearer invalid_token"}
+    )
+
+    assert response.status_code == 401
+    assert response.json()["detail"] == "Invalid token"
+
+def test_post_ingredient_category_without_token(client: TestClient):
+    response = client.post(f"{url_prefix}/", 
+        json={
+            "name": "test ingredient category",
+            "created_by": ADMIN_ID
+        }
+    )
+
+    assert response.status_code == 401
+    assert response.json()["detail"] == "Not authenticated"
+
+def test_post_ingredient_category_with_various_letter_case(client: TestClient, token: str):
+    response = client.post(f"{url_prefix}/", 
+        json={
+            "name": "Test INgrEDient catEGory",
+            "created_by": ADMIN_ID
+        },
+        headers={"Authorization": f"Bearer {token}"}
+    )
+
+    assert response.status_code == 201
+    assert response.json()["detail"] == f"test ingredient category as Ingredient Category is successfully created"
+
+    ingredient_category = response.json()["ingredient_category"]
+
+    assert ingredient_category["name"] == "test ingredient category"
+    assert ingredient_category["creator"]["id"] == "db67b3f4-0e04-47bb-bc46-94826847ee4f"
+    assert "updated_date" in ingredient_category
+    assert "created_date" in ingredient_category
+
+def test_post_ingredient_category_with_duplicate_name(client: TestClient, token: str):
+    response = client.post(f"{url_prefix}/", 
+        json={
+            "name": "proteins",
+            "created_by": ADMIN_ID
+        },
+        headers={"Authorization": f"Bearer {token}"}
+    )
 
     assert response.status_code == 400
     assert response.json()["detail"] == f"proteins as Ingredient Category is already registered"
 
-def test_post_ingredient_category_with_empty_name(client: TestClient):
-    response = client.post(f"{url_prefix}/", json={
-        "name": "",
-        "created_by": ADMIN_ID
-    })
+def test_post_ingredient_category_with_empty_name(client: TestClient, token: str):
+    response = client.post(f"{url_prefix}/", 
+        json={
+            "name": "",
+            "created_by": ADMIN_ID
+        },
+        headers={"Authorization": f"Bearer {token}"}
+    )
 
     assert response.status_code == 422
 
@@ -89,10 +174,13 @@ def test_post_ingredient_category_with_empty_name(client: TestClient):
     assert response_json["detail"][0]["msg"] == "String should have at least 3 characters"
     assert response_json["detail"][0]["type"] == "string_too_short"
     
-def test_post_ingredient_category_without_name(client: TestClient):
-    response = client.post(f"{url_prefix}/", json={
-        "created_by": ADMIN_ID
-    })
+def test_post_ingredient_category_without_name(client: TestClient, token: str):
+    response = client.post(f"{url_prefix}/", 
+        json={
+            "created_by": ADMIN_ID
+        },
+        headers={"Authorization": f"Bearer {token}"}
+    )
 
     assert response.status_code == 422
 
@@ -105,20 +193,26 @@ def test_post_ingredient_category_without_name(client: TestClient):
     assert response_json["detail"][0]["msg"] == "Field required"
     assert response_json["detail"][0]["type"] == "missing"
 
-def test_post_ingredient_category_with_not_available_creator_id(client: TestClient):
-    response = client.post(f"{url_prefix}/", json={
-        "name": "test ingredient category",
-        "created_by": "3fa85f64-5717-4562-b3fc-2c963f66afa6"
-    })
+def test_post_ingredient_category_with_not_available_creator_id(client: TestClient, token: str):
+    response = client.post(f"{url_prefix}/", 
+        json={
+            "name": "test ingredient category",
+            "created_by": "3fa85f64-5717-4562-b3fc-2c963f66afa6"
+        },
+        headers={"Authorization": f"Bearer {token}"}
+    )
 
     assert response.status_code == 404
     assert response.json()["detail"] == f"Id 3fa85f64-5717-4562-b3fc-2c963f66afa6 as User is not found"
 
-def test_post_ingredient_category_with_invalid_creator_id(client: TestClient):
-    response = client.post(f"{url_prefix}/", json={
-        "name": "test ingredient category",
-        "created_by": "3fa85f64-5717-4562-b3fc-2c963f66afaz"
-    })
+def test_post_ingredient_category_with_invalid_creator_id(client: TestClient, token: str):
+    response = client.post(f"{url_prefix}/", 
+        json={
+            "name": "test ingredient category",
+            "created_by": "3fa85f64-5717-4562-b3fc-2c963f66afaz"
+        },
+        headers={"Authorization": f"Bearer {token}"}
+    )
 
     assert response.status_code == 422
 
@@ -131,11 +225,14 @@ def test_post_ingredient_category_with_invalid_creator_id(client: TestClient):
     assert response_json["detail"][0]["msg"] == "Input should be a valid UUID, invalid character: expected an optional prefix of `urn:uuid:` followed by [0-9a-fA-F-], found `z` at 36"
     assert response_json["detail"][0]["type"] == "uuid_parsing"
 
-def test_post_ingredient_category_with_empty_creator_id(client: TestClient):
-    response = client.post(f"{url_prefix}/", json={
-        "name": "test ingredient category",
-        "created_by": ""
-    })
+def test_post_ingredient_category_with_empty_creator_id(client: TestClient, token: str):
+    response = client.post(f"{url_prefix}/", 
+        json={
+            "name": "test ingredient category",
+            "created_by": ""
+        },
+        headers={"Authorization": f"Bearer {token}"}
+    )
 
     assert response.status_code == 422
 
@@ -148,10 +245,13 @@ def test_post_ingredient_category_with_empty_creator_id(client: TestClient):
     assert response_json["detail"][0]["msg"] == "Input should be a valid UUID, invalid length: expected length 32 for simple format, found 0"
     assert response_json["detail"][0]["type"] == "uuid_parsing"
 
-def test_put_ingredient_category(client: TestClient):
-    response = client.put(f"{url_prefix}/b4b165f6-a4f2-45f6-bda6-0a49092d3f03", json={
-        "name": "updated ingredient category",
-    })
+def test_put_ingredient_category_by_changing_name(client: TestClient, token:str):
+    response = client.put(f"{url_prefix}/b4b165f6-a4f2-45f6-bda6-0a49092d3f03", 
+        json={
+            "name": "updated ingredient category",
+        },
+        headers={"Authorization": f"Bearer {token}"}
+    )
 
     assert response.status_code == 202
     assert response.json()["detail"] == "Id b4b165f6-a4f2-45f6-bda6-0a49092d3f03 as Ingredient Category is successfully updated"
@@ -164,18 +264,64 @@ def test_put_ingredient_category(client: TestClient):
     assert "updated_date" in ingredient_category
     assert "created_date" in ingredient_category
 
-def test_put_ingredient_category_with_duplicate_name(client: TestClient):
-    response = client.put(f"{url_prefix}/b4b165f6-a4f2-45f6-bda6-0a49092d3f03", json={
-        "name": "vegetables",
-    })
+def test_put_ingredient_category_by_changing_name_with_invalid_token(client: TestClient):
+    response = client.put(f"{url_prefix}/b4b165f6-a4f2-45f6-bda6-0a49092d3f03", 
+        json={
+            "name": "updated ingredient category",
+        },
+        headers={"Authorization": f"Bearer invalid_token"}
+    )
+
+    assert response.status_code == 401
+    assert response.json()["detail"] == "Invalid token"
+
+def test_put_ingredient_category_by_changing_name_without_token(client: TestClient, token:str):
+    response = client.put(f"{url_prefix}/b4b165f6-a4f2-45f6-bda6-0a49092d3f03", 
+        json={
+            "name": "updated ingredient category",
+        }
+    )
+
+    assert response.status_code == 401
+    assert response.json()["detail"] == "Not authenticated"
+
+def test_put_ingredient_category_by_changing_name_with_various_letter_case(client: TestClient, token:str):
+    response = client.put(f"{url_prefix}/b4b165f6-a4f2-45f6-bda6-0a49092d3f03", 
+        json={
+            "name": "upDATed inGREdient cATEgory",
+        },
+        headers={"Authorization": f"Bearer {token}"}
+    )
+
+    assert response.status_code == 202
+    assert response.json()["detail"] == "Id b4b165f6-a4f2-45f6-bda6-0a49092d3f03 as Ingredient Category is successfully updated"
+
+    ingredient_category = response.json()["ingredient_category"]
+
+    assert ingredient_category["id"] == "b4b165f6-a4f2-45f6-bda6-0a49092d3f03"
+    assert ingredient_category["name"] == "updated ingredient category"
+    assert ingredient_category["creator"]["id"] == "db67b3f4-0e04-47bb-bc46-94826847ee4f"
+    assert "updated_date" in ingredient_category
+    assert "created_date" in ingredient_category
+
+def test_put_ingredient_category_by_changing_name_with_duplicate_name(client: TestClient, token:str):
+    response = client.put(f"{url_prefix}/b4b165f6-a4f2-45f6-bda6-0a49092d3f03", 
+        json={
+            "name": "vegetables",
+        },
+        headers={"Authorization": f"Bearer {token}"}
+    )
 
     assert response.status_code == 400
     assert response.json()["detail"] == f"'vegetables' as Ingredient Category is already registered"
 
-def test_put_ingredient_category_with_empty_name(client: TestClient):
-    response = client.put(f"{url_prefix}/b4b165f6-a4f2-45f6-bda6-0a49092d3f03", json={
-        "name": "",
-    })
+def test_put_ingredient_category_with_empty_name(client: TestClient, token:str):
+    response = client.put(f"{url_prefix}/b4b165f6-a4f2-45f6-bda6-0a49092d3f03", 
+        json={
+            "name": "",
+        },
+        headers={"Authorization": f"Bearer {token}"}
+    )
 
     assert response.status_code == 422
 
@@ -188,18 +334,24 @@ def test_put_ingredient_category_with_empty_name(client: TestClient):
     assert response_json["detail"][0]["msg"] == "String should have at least 3 characters"
     assert response_json["detail"][0]["type"] == "string_too_short"
 
-def test_put_ingredient_category_with_not_available_creator_id(client: TestClient):
-    response = client.put(f"{url_prefix}/b4b165f6-a4f2-45f6-bda6-0a49092d3f03", json={
-        "created_by": "3fa85f64-5717-4562-b3fc-2c963f66afa6"
-    })
+def test_put_ingredient_category_with_not_available_creator_id(client: TestClient, token:str):
+    response = client.put(f"{url_prefix}/b4b165f6-a4f2-45f6-bda6-0a49092d3f03", 
+        json={
+            "created_by": "3fa85f64-5717-4562-b3fc-2c963f66afa6"
+        },
+        headers={"Authorization": f"Bearer {token}"}
+    )
 
     assert response.status_code == 404
     assert response.json()["detail"] == f"Id 3fa85f64-5717-4562-b3fc-2c963f66afa6 as User is not found"
 
-def test_put_ingredient_category_with_invalid_creator_id(client: TestClient):
-    response = client.put(f"{url_prefix}/b4b165f6-a4f2-45f6-bda6-0a49092d3f03", json={
-        "created_by": "3fa85f64-5717-4562-b3fc-2c963f66afaz"
-    })
+def test_put_ingredient_category_with_invalid_creator_id(client: TestClient, token:str):
+    response = client.put(f"{url_prefix}/b4b165f6-a4f2-45f6-bda6-0a49092d3f03", 
+        json={
+            "created_by": "3fa85f64-5717-4562-b3fc-2c963f66afaz"
+        },
+        headers={"Authorization": f"Bearer {token}"}
+    )
 
     assert response.status_code == 422
 
@@ -212,10 +364,13 @@ def test_put_ingredient_category_with_invalid_creator_id(client: TestClient):
     assert response_json["detail"][0]["msg"] == "Input should be a valid UUID, invalid character: expected an optional prefix of `urn:uuid:` followed by [0-9a-fA-F-], found `z` at 36"
     assert response_json["detail"][0]["type"] == "uuid_parsing"
 
-def test_put_ingredient_category_with_empty_creator_id(client: TestClient):
-    response = client.put(f"{url_prefix}/b4b165f6-a4f2-45f6-bda6-0a49092d3f03", json={
-        "created_by": ""
-    })
+def test_put_ingredient_category_with_empty_creator_id(client: TestClient, token:str):
+    response = client.put(f"{url_prefix}/b4b165f6-a4f2-45f6-bda6-0a49092d3f03", 
+        json={
+            "created_by": ""
+        },
+        headers={"Authorization": f"Bearer {token}"}
+    )
 
     assert response.status_code == 422
 
@@ -228,14 +383,26 @@ def test_put_ingredient_category_with_empty_creator_id(client: TestClient):
     assert response_json["detail"][0]["msg"] == "Input should be a valid UUID, invalid length: expected length 32 for simple format, found 0"
     assert response_json["detail"][0]["type"] == "uuid_parsing"
 
-def test_delete_ingredient_category(client: TestClient):
-    response = client.delete(f"{url_prefix}/b4b165f6-a4f2-45f6-bda6-0a49092d3f03")
+def test_delete_ingredient_category(client: TestClient, token: str):
+    response = client.delete(f"{url_prefix}/b4b165f6-a4f2-45f6-bda6-0a49092d3f03", headers={"Authorization": f"Bearer {token}"})
     
     assert response.status_code == 200
     assert response.json()["detail"] == "Id b4b165f6-a4f2-45f6-bda6-0a49092d3f03 as Ingredient Category is successfully deleted"
 
-def test_delete_ingredient_category_by_wrong_id(client: TestClient):
-    response = client.delete(f"{url_prefix}/db67b3f4-0e04-47bb-bc46-94826847ee4f")
+def test_delete_ingredient_category_with_invalid_token(client: TestClient):
+    response = client.delete(f"{url_prefix}/b4b165f6-a4f2-45f6-bda6-0a49092d3f03", headers={"Authorization": f"Bearer invalid_token"})
+
+    assert response.status_code == 401
+    assert response.json()["detail"] == "Invalid token"
+
+def test_delete_ingredient_category_without_token(client: TestClient):
+    response = client.delete(f"{url_prefix}/b4b165f6-a4f2-45f6-bda6-0a49092d3f03")
+
+    assert response.status_code == 401
+    assert response.json()["detail"] == "Not authenticated"
+
+def test_delete_ingredient_category_by_wrong_id(client: TestClient, token: str):
+    response = client.delete(f"{url_prefix}/db67b3f4-0e04-47bb-bc46-94826847ee4f", headers={"Authorization": f"Bearer {token}"})
 
     assert response.status_code == 404
     assert response.json()["detail"] == "Id db67b3f4-0e04-47bb-bc46-94826847ee4f as Ingredient Category is not found"
